WasmThemis utilities (#489)
* Correctly report error codes

Classical off-by-one mistake. We incorrectly reported error codes as
'undefined' when doing "new ThemisError(subsystem, errorCode)". This
affects only human-readable description, not the 'errorCode' property,
but it's very annoying for developers.

Use ThemisError.describe only when there are 3 (three) or more arguments
with human-readable error causes.

* Move 'coerceToBytes' into utility module

This function is going to be useful in multiple places for type checking
of user byte arrays. Let's move it into a common utility module. While
we're here, let it throw 'TypeError' instead of 'ThemisError' which
should be more recognizable (and does not expect Themis "subsystem").