Secure Cell support in WasmThemis (#462)
Introduce Secure Cell support starting with its sealing mode.

* Secure Cell: Seal mode

We keep the same structure as in JsThemis: the main module exports three
classes: SecureCellSeal, SecureCellTokenProtect, SecureCellContextImprint.
All of them have a constructor that accepts a master key, all of them
have "encrypt" and "decrypt" methods which accept a message and an
optional context.

Implementation involves quite a lot of copy-pasta so the individulal
modes have separate source files. Unfortunately, I was not able to
reduce code duplication much because of the subtle differences between
modes. Therefore the code is kept mostly dumb, involving various FFI
things we need to do to transfer the data into Emscripten heap and back.

Sealing mode is the easiest one and the most straightforward. Notable
things are processing of optional context arguments, overall interfacing
with C, and zeroing out of the sensitive master key after we're done
with it. Unfortunately, JavaScript does not have RAII and we have to
copy the data to Emscripten heap every time we need to call C code.

* Secure Cell: Context Imprint mode

It's similar to the sealing mode, but the 'context' is required for
Context Imprint mode (duh...) We reflect this by producing a more
helpful error message if the user forgets to call encrypt/decrypt with
two arguments. The test set is also a bit different. And note that the
argument order for C functions is slightly different as well. Other that
that, it's mostly a copy of the sealing mode.

* Secure Cell: Token Protect mode

Token Protect mode is different in that it returns *two* values: the
message and the authentication token separately. JavaScript does not
have tuples so we return an object with "data" and "token" fields,
just like JsThemis. For decryption we require data and token to be
provided as separate arguments. This is a bit weird for tests, but
should not cause issues in real-world usage.

* Allow null context in Secure Cell

Treat "null" context value as omitted or empty context for Sealing and
Token Protect modes. This provides more natural interface for users.

Note that Context Imprint mode requires non-empty context thus it still
treats null values as errors.