Do not call KDF during size query (decryption)
themis_auth_sym_decrypt_message() has the same inefficient use of
themis_sym_kdf() as in themis_auth_sym_encrypt_message(). Update
and refactor it in a similar way.

Complexity here stems from the fact that decryption path contains
a workaround for a compatibility issue with Themis 0.9.6. We need
to preserve the workaround. In order to make it more readable,
move the compatibility code into a separate function that will be
compiled out entirely when not needed.