Always fill buffer length in Secure Cell (#497)
Sealing mode of Secure Cell is a bit special in that its output is
a concatenation of encrypted message and authentication context.
Other modes return these components separately.

The users need to allocate a suitably big buffer to fit the results of
encryption.  Usually they pass NULL for buffer pointer in order to
measure the size first, then allocate the buffer and pass the correct
buffer pointer and size to Themis.

However, the users might also allocate a big enough buffer right away
and do not rely on measurements. Or allocate a buffer with some extra
rounding (e.g., to a page size or something). In this case they need
to know the resulting length too.

Unforuntately, Secure Cell in sealing mode does not fill in the inout
length parameter if the buffer turns out to be big enough already.
This will result in extra unused and uninitialized content at the end
of the encrypted message if the user decides to trust the length.

Make sure that we return the length in this case. Add a test to check
that resulting length is consistent with prior measurements. Check
other modes as well, just in case.