Install RVM from get.rvm.io
Unfortunately, latest updates to RVM installer has broken the PPA
package that we are using to install RVM into the system. Versions
1.29.9+ are broken when installing in environment with sudo so we
can't use it until this issue is fixed.

Instead of installing RVM into the system in multi-user mode, install
RVM for the single user as instructed by RVM docs. Make sure to select
system Ruby installation after that.