Redundant key server pools (#504)
* Redundant key server pools

Using just "pool.sks-keyservers.net" as recommended by RVM documentation
seems to be too unreliable for CI setup. It works fine for singular
installations, but when deployed to CircleCI pool-selected servers are
often unavailable, time out, and break our builds. Instead of using a
single server try multiple ones for each key that we wish to request.

* Mention source for key IDs and key servers

This makes it easier to verify key IDs and update them as necessary
if importing the keys or the signature checks that follow start failing.