Fix error handling in key generation (#500)
Note that EC key generation has incorrect return value when public key
export fails. This may cause key generation function to return success
if private key has been exported successfully, but public key has not.
The user will then proceed with uninitalized public key buffer and will
not be amused afterwards.

Change variable names to be more distinct so that the difference is more
obvious and add some empty lines so that the code is easier to read.

Add a test which makes sure we have correct behavior when one of the
keys fail. Using NULL values is the easiest way to trigger this
behavior.