Test JsThemis with actual Node.js versions (#502)
* Workflow for testing JsThemis with multiple Node.js

It turns out that it's not enough to test JsThemis with Node.js
that is bundled with the distribution we use (Ubuntu 16.04 LTS).
This version is quite old and not really supported by Joyent.
We should be testing with the modern version as well. Currently
these are v8 (old LTS), v10 (current LTS), and v12 (next LTS).

Add a dedicated workflow which will build and test JsThemis
across these versions of Node.js. We use the NVM tool to manage
Node.js installations.

* Remove JsThemis tests from x86_64 workflow

Since JsThemis is now tested in its dedicated workflow there is no need
to install Node.js and test JsThemis in the common x86_64 workflow.
Remove this redundancy