Package ThemisPP
Introduce DEB and RPM packages for ThemisPP. Since C++ is system
language related to C, it's natural to provide our C++ wrapper as
a system package. The wrapper requires Themis Core to be installed
so this plays out nicely in the end.

Alternatively, we can provide a package for various other package
managers for C++ (such as Conan, vcpkg, CMake, Meson, and a zillion
others) but there isn't much demand for that, and we'd have to provide
Themis Core there as well for proper versioning and dependencies.
So let's not complicate it more than necessary.

Since ThemisPP is header-only we provide development packages:
"libthemispp-dev" and "libthemispp-devel". There is no "libthemispp"
package that's required at runtime: just "libthemis" is enough.

While we're here, cleanup some name duplication in package names and
dependencies.