Makefile targets for WasmThemis installation
Provide some utility targets to install and uninstall WasmThemis into
current directory. These are not really usefult for our users, but they
will make CI's life a bit easier.

Rename the main WasmThemis target to "wasmthemis" in order to be
consinstent with installation names.

Note that we use absolute path to the tarball because "npm install" can
be easily confused with relative paths like "build/wasm-themis.tgz"
which is a valid name for a scoped npm package.