Prepare for publishing WasmThemis (#513)
* Remove Browserify reference from README

This is not the only bundler in existance. We're going to discuss
bundling in the documentation so we don't need this here. Supposedly,
Web developers know what to do with npm packages to get them shipped.

* Update package.json for wasm-themis

Remove "private" option so that the package can be published in public
npm repositories.

Update "homepage" to an actual home page of Themis, not just README
which is reachable via the repository link.

* Update .npmignore

We have moved some of the internal build files but did not update
.npmignore so that they won't get packaged in the release build.