Enable more clang-tidy checks
We've been using not that that much checks. Let's enable whatever
available. Note, however, that code base has quite a few legacy
warts so we can't enable everything and have to disable some of
the warnings.

We still support C++03 so we do not enable "modernize-*" group.
This hides an amazing amount of warnings from JsThemis code
(which ended up with its own clang-tidy configuration).

Some other tweaks:

- readability-implicit-bool-conversion - because we have quite
  a few places where we check integers for zero with "!" and
  all of them trigger this warning

- readability-isolate-declaration - "sput" (our testing framework)
  macros cause most of these warnings, hopefully we can enable
  this back after we migrate to a newer framework for unit tests

- *-magic-numbers - unfortunately, clang-tidy gives a warning
  for *every individual integer* in data arrays (like in CRC32
  code) which is idiotic, sorry but this goes out of the window

- cppcoreguidelines-pro-type-vararg - we use printf() functions
  quite often (due to sput and SOTER_IF macros), don't want this
  to be triggered. Maybe we can disable this later