Correct EVP_PKEY ownership: soter_rsa_key_pair_gen
OpenSSL and BoringSSL implementations are slightly different. Currently
we don't have time to dive deep into the reasons behind this so we just
fix the ownership issues here and move on.

In this case it makes sense to introduce a couple of utility functions
to hide some complexity of dealing with BIGNUMs.