OpenSSL compatibility shims
There are a couple of warnings about unused labels. They are caused
by the old compatibility code not using the labels because it does
not have failure conditions that we present in the new code.

Instead of using ifdefs in the code base, introduce compatibility
shims for newer functions like RSA_get0_key() that we not available
in older OpenSSL versions. This forces us to handle possible error
conditions and gets rid of the warnings.

Additionally, all compatibility code is now compactly located in
one place and can be easily removed later.