Suppress warnings about CRYPTO_ENGINE_PATH
For some historical reasons signature checking code needs access to
private data structures of cryptographic backend. We achieve that
be cleverly applying advanced preprocessor macrology. Unfortunately,
it confuses clang-tidy which emits a lint here. Suppress the lint
and avoid code duplication.

These changes are limited to "soter_t.h" which is a private header file.
They are not visible to the users, but we should review the code later
and try to avoid dependency on the implementation details like the
choice of cryptographic backend.