Improve memory safety of tests
Static analyzer gets confused by all possible code paths through this
enormous function, and produces a bunch of warnings about possible NULL
pointer dereference and double-free errors. Let's help it a little by
making sure that we handle NULL values in the helper functions and
by zeroing out recently freed pointers.

Yeah, it would be better to refactor this complex functions, but let's
keep it as is for now. We'll update it later, together with a major
unit-test overhaul.

(Still, I'm amazed at Clang being able to see a possible error condition
which requires around a dozen of other conditions to hold in order for
it to have a chance at triggering.)