Fix warnings about possible ODR violations
Non-inline and non-static functions in header files may trigger
weird issues with one-definition rule (ODR) if the definitions
are different. It's obviously not true in our case but you tell
that to clang-tidy.

Mark private functions as static and exported functions as inline
to avoid these warnings.

Ideally, we'd put the actual code to *.cpp files. We'll do that
later together with proper refactoring of unit tests.