Fix rule-of-five violations
First of all, add explicit virtual destructors to classes that declare
virtual functions. This is important for deallocation to work properly,
especially for the callback interface of Secure Session. (Other classes
do not really need inheritance, but they currently use it and thus need
the destructors as well.)

After that, bring back copy constructor to secure_message_t. There iso
no reason for these objects to not be copyable or moveable.

Finally, implement a move constructor for secure_comparator_t. These
objects can be moved, but cannot be copied. Add a move constructor
and move-assignment operator for them. Delete the copy constructor
and copy-assignment operator.

Note that we keep compatibility with C++03 and provide explicit empty
functions and the 'private constructor' hack there. With C++11 we can
use default and deleted implementations directly.