Use vectors for data instead of strings
Various code snipptes like

    std::vector<uint8_t> pass1(password1.c_str(),
                               password1.c_str() + password1.length());

cause clang-tidy warnings about pointer arithmetic being dangerous and
discouraged. ThemisPP interfaces work with std::vector<uint8_t>, so
instead of using std::string, let's use std::vector directly.

Introduce a helper function to convert a string into a vector,
replace all inline instances of such conversion with just references
to a global constant with test data. This fixes most of the warnings.

There is more interesting story with Secure Session tests. First of
all, it requires private keys which are obvioulsy not strings. Currently
we use C arrays which get two warnings: one about using C arrays
(in C++11 there is std::array which is safer) and another one for
inevitable use of pointer arithmetic.

In C++11 we'd use initialization lists to construct a vector from
C array syntax, but C++03 does not have that option so we have to
keep the C arrays and convert them into vectors.

Another thing is replacing std::string with std::vector as a key for
std::unordered_map that we use to keep client IDs. The thing is that
std::vector does not have a built-in implementation of std::hash
so we have to provide our own for that code to compile.