Avoid C-style pointer casts
There is another warning for C-style pointer casts. In this case we
know exactly what kind of pointer is there under the disguise of void*
so we can use static_cast instead.

Also, mark pointer arithmetic when constructing vectors with NOLINT
since we really can't get around that: that's C callback interface.