Fix pass-by-reference API
This method is missing a reference symbol, most likely due to a typo.
Make it accept a reference which in turns makes the warning about
useless "const" specifier to go away.

Also, this technically changes the ABI, but since C++ does not have
a stable ABI, it should be okay. Previously written code will be
okay with this change.