Remove unnecessary const specifiers
clang-tidy produces a warning about this case. There is really
little use to return a const vector by copy here. Drop that.

Unfortunately, this changes the API and ABI of the code and user
applications may need to be updated to compile against the newer
version. It is not possible to degrade this gracefully. Though,
the change is minor and should not cause issues when upgrading.