Hide public field that's private
There is no real reason for this field to be public. It's been
like that since the beginning.

Instead of exposing the field as public, declare the function that
needs access to it as private static method of SecureSession class.
This provides access to private members and we can still use this
function as a callback.