Roll back to Emscripten 1.38.48 (#550)
Recently released 1.39.0 seems to have a regression than breaks our
build. Let's roll back to the previous version until the issue is
examined and resolved.

(cherry picked from commit c2fcdc5998510356348f23ea57510113ea7c9680)

A-a-and we also need this so that CircleCI uses correct Emscripten
flavor for running tests.