Link libthemis_jni.so dynamically
Do not embed Themis, Soter, and cryptographic backend (if present)
into libthemis_jni shared object. Instead, link it dynamically
against Themis like a proper library.

It has been a nice experiment to provide self-contained library for
Java. However, we are going to distribute the library as a proper system
package and for that we will need to have it properly linked against its
dependencies. Themis Core will be installed by the package manager so
all dependencies are going to be satisfied.