Installation targets for libthemis_jni
Introduce two new Makefile targets to install and uninstall Themis JNI
library: themis_jni_{install,uninstall}

The library is installed to $(jnidir) which is set to $(libdir) by
default, but can be overriden. This is because Java is not required
to look into standard system directories for JNI libraries and may
have a different opinion about that.

Note that libthemis_jni does not have ABI suffix, does not introduce
development symbolic links, and does not define import libraries like
Soter and Themis libraries do. This is because it is not meant to be
used by linkers, only dynamically loaded by JVM. Therefore we can
simply name it "libthemis_jni.so" and that's all: it's going to be
loaded by path anyway, its SONAME does not matter.

Also note that on macOS we invoke install_name_tool only during
installation, not during build. This is in-line with other libraries
and is the correct behavior (so that we do not have to rebuild the
library before installing it into a different place).