Warn the user about path mismatch
Let's check Java library search paths and print out a warning if JNI
library is being installed to a locations that's probably unreachable
by default.

Java installations are not required to look into system library
directories. While normally OpenJDK and Oracle JDK installations
do look into some directory in the /usr hierarchy, *none* of the systems
that we support looks into "/usr/local/lib" where we install by default
when compiling from source.

This is unfortunate, but we should not decide for the user how they
should be configuring their system when building from source, and we
should not interfere with system package manager either.

For Linux we provide packages that install JNI library into /usr
subdirectories expected by Java. This is the preferred way to install
Themis Core on Linux. Users that build from source will have to adjust
JNI library location manually after installation.

Unfortunately, on macOS Homebrew is not able to write to any of the
Oracle's Java default paths except for the one in user's home directory:

    $HOME/Library/Java/Extensions
    /Library/Java/Extensions         - requires root acccess
    /Network/Library/Java/Extensions - protected by SIP
    /System/Library/Java/Extensions  - protected by SIP
    /usr/lib/java                    - protected by SIP
    $PWD

Installing into $HOME or the current directory is not really an option
for Homebrew. Thus on macOS we are most likely installing the library
into a directory that's not reachable by default. Homebrew package
should be adjusted to print out a relevant message to the user.