Don't suppress UBSan unless compiling with UBSan (#556)
Recently added SOTER_ED25519_NO_UBSAN macro is expanded into UBSan
tweaks even when we do not compile with Undefined Behavior sanitizer
enabled. This can produce ugly warnings on systems that do not support
all of the sanitizer flags.

Let's just expand this macro into a no-op if we are not compiling with
sanitizers. That's easier that deducing exactly which sanitizer flags
are supported to avoid compiler warnings.