Fix Secure Comparator in PyThemis on 32-bit (#555)
secure_comparator_get_result() returns themis_status_t which is defined
as int32_t. We should expect 32-bit integer to be returned here.

This kinda worked on 64-bit platforms because of how registers work on
x86_64 and how 64-bit values are returned in System V ABI. However, this
does not work on i686 and causes Python to see garbage returned.