Migrate Java tests to JUnit4
New Android instrumentation testing framework deprecates AndroidTestCase
in favor of more common JUnit4 framework. Convert the tests to use that.
It's a quick-and-dirty conversion to just make it run. We'll improve the
quality of tests later.

As a side effect, this allows us to reuse those tests as unit tests, not
only for Android, but for desktop Java as well.