Update and enable Android on CircleCI
Use a new Docker image for running tests. It contains updated Android
SDK and NDK, platform tools, system images, etc.

Remove the concurrency hacks since they seem to be no longer necessary
on CirceCI. We can avoid tweaking the build files and explicitly
specifying "--no-parallel --max-workers=2" in Gradle invocations.

Update the system image used for testing. Unfortunately, we still
need to use ARM emulation because x86 and x86_64 requires KVM which is
not available on CircleCI runners. More unfortunately, we need at least
Android API 24 in order to run new tests, and Android 24 is the last
version that has ARM images. It looks like we'll get into a pinch
when Google eventually decides to deprecate and remove the image.

Tweak the command line for running the emulator. Previously we had
disabled GPU, but that seems to break the current emulator and it gets
stuck in the boot process without going anywhere. Therefore, run with
GPU enabled. (We still don't need and cannot open a window though.)

All in all, this allows us to finally reenable Android build job!